By default compiles/install both sbcl and ccl
Only tries to install versions for which the required lisp is found in
the user's path.  Different executables are compiled to buildapp-sbcl
and buildapp-ccl respectively.  The buildapp file is a short shell
script which dispatches to the appropriate buildapp version depending on
the value of the LISP environment variable (defaulting to SBCL).

See the diff of doc/index.html for additional information.